requirements.txt
timm==0.6.7
torch==2.0.1
torchvision==0.15.2
numpy==1.23.5
pillow==9.4.0
gradio==4.18.0
opencv-python==4.6.0.66
kaleido==0.2.1